Rename input error helper
