+++ src/components/form/TextInput.tsx
@@ -9,11 +9,13 @@ import { TextInput as RNTextInput } from "react-native";
 import { TextInput as RNPTextInput } from "react-native-paper";
 
 // Components
-import InputError from "./InputError";
+import InputHelperText from "./InputHelperText";
 
 type TextInputProps = {
   /** Control from 'react-hook-form' */
   control: Control<any>;
+  /** Hint text */
+  hint?: string;
   /** TextInput ref */
   innerRef?: RefObject<RNTextInput>;
   /** Form control name */
@@ -21,7 +23,7 @@ type TextInputProps = {
 } & ComponentProps<typeof RNPTextInput>;
 
 const TextInput = (props: TextInputProps): ReactElement => {
-  const { control, innerRef, name, ...rest } = props;
+  const { control, hint, innerRef, name, ...rest } = props;
 
   const { field, fieldState, formState } = useController({
     control,
@@ -43,7 +45,11 @@ const TextInput = (props: TextInputProps): ReactElement => {
         onBlur={field.onBlur}
         onChangeText={field.onChange}
       />
-      <InputError error={error?.message} visible={errorShown} />
+      <InputHelperText
+        error={error?.message}
+        hint={hint}
+        visible={Boolean(errorShown || hint)}
+      />
     </Fragment>
   );
 };
