+++ src/localization/en/common.json
@@ -21,6 +21,7 @@
   },
   "phrases": {
     "comingSoon": "Coming Soon!",
+    "optional": "Optional",
     "search": "Search"
   },
   "validations": {
