+++ src/localization/es/common.json
@@ -21,6 +21,7 @@
   },
   "phrases": {
     "comingSoon": "Próximamente!",
+    "optional": "Opcional",
     "search": "Buscar"
   },
   "validations": {
